test(epub): extract metadata profile family
