+++ src/research/epub-test-utils.ts
@@ -1,0 +1,14 @@
+import { zipSync } from 'fflate'
+
+function rezipEpub(files: Record<string, Uint8Array>) {
+  return zipSync({
+    mimetype: [files.mimetype, { level: 0 }],
+    ...Object.fromEntries(
+      Object.entries(files)
+        .filter(([name]) => name !== 'mimetype')
+        .map(([name, bytes]) => [name, [bytes, { level: 6 }]]),
+    ),
+  })
+}
+
+export { rezipEpub }
